// src/generated/models/Cr22fProjetoModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface Cr22fProjetoBase {
  cr22f_projetoid: string;
  cr22f_ano?: string;
  cr22f_apelido?: string;
  cr22f_url?: string;
  "new_Arquiteto@odata.bind"?: string;
  "new_Cliente@odata.bind"?: string;
  "new_Coordenador@odata.bind"?: string;
  "new_Projetista@odata.bind"?: string;
  statecode?: number;
  statuscode?: number;
}

export interface Cr22fProjeto extends Cr22fProjetoBase {
  _new_arquiteto_value?: string;
  _new_cliente_value?: string;
  _new_coordenador_value?: string;
  _new_projetista_value?: string;
  new_arquiteto_name?: string;
  new_cliente_name?: string;
  new_coordenador_name?: string;
  new_projetista_name?: string;
}