// src/generated/models/NewS3objectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface NewS3objectsBase {
  new_s3objectsid: string;
  new_bucketname?: string;
  new_datetime?: string;
  new_id?: string;
  new_isimage?: boolean;
  new_localurl?: string;
  new_nomedocliente?: string;
  new_nomedoprojeto?: string;
  new_nomedoremetente?: string;
  new_remoteurl?: string;
  new_wazzupchatid?: string;
  new_wazzupmessagedatetime?: string;
  new_wazzupmessagetext?: string;
  "new_ComponenteDescricaoItemCatalogoVerificacao@odata.bind"?: string;
  "new_FeedbackdoColaborador@odata.bind"?: string;
  "new_OrdemdeServico@odata.bind"?: string;
  "new_RMA@odata.bind"?: string;
  "new_UserLocation@odata.bind"?: string;
  "new_WazzupMessage@odata.bind"?: string;
  statecode?: number;
  statuscode?: number;
}

export interface NewS3objects extends NewS3objectsBase {
  _new_componentedescricaoitemcatalogoverificacao_value?: string;
  _new_feedbackdocolaborador_value?: string;
  _new_ordemdeservico_value?: string;
  _new_rma_value?: string;
  _new_userlocation_value?: string;
  _new_wazzupmessage_value?: string;
}